Second Review Amendments
- Updated README.
- Added convenience script to launch test server.
- Added removal of all accounts on launch when UI_TESTING argument is detected.
- Removed redundant code. This might be a rebase accident.
- Removed obsolete shell scripts for test servers.
- Consolidated XCUIElement extensions into dedicated source code group a single file.
