+++ .github/workflows/xcode.yml
@@ -5,7 +5,6 @@ on:
     branches:
       - master
       - develop
-      - 3184-download-limit2 # ONLY TEMPORARILY TO VERIFY THE CHANGES IN HERE
   pull_request:
     types:
       - synchronize
@@ -69,7 +68,6 @@ jobs:
         extensions: apcu, bz2, ctype, curl, dom, fileinfo, gd, iconv, intl, json, libxml, mbstring, openssl, pcntl, posix, session, simplexml, xmlreader, xmlwriter, zip, zlib, pgsql, pdo_pgsql
         coverage: none
         ini-file: development
-        # Temporary workaround for missing pcntl_* in PHP 8.3: ini-values: apc.enable_cli=on
         ini-values: apc.enable_cli=on, disable_functions=
 
     - name: Checkout Nextcloud Server
