+++ README.md
@@ -78,44 +78,25 @@ Do you want to try the latest version in development of Nextcloud iOS ? Simple, 
 
 ### Unit Tests
 
-There are currently no prerequisites for unit testing. Mock everything which not needed. 
+There are currently no prerequisites for unit testing.
+Mock everything which is not needed. 
 
 ### Integration Tests
 
 In `TestConstants.swift` you must specify your credentials for the server you want to test against.
-[The Nextcloud Docker image](https://github.com/nextcloud/docker) maintained by the community is an easy and fast way to spin up a Nextcloud server to test against. 
-[This shallow CI server image](https://github.com/nextcloud/docker-ci/pkgs/container/continuous-integration-shallow-server) is also used by the Android project for automated tests in the GitHub pipeline.
-This latter one will be referred to from here on.
-With the following commands, you have will set up a new Nextcloud available on local HTTP port 8080, once the Docker image was initially pulled.
+For locally running tests and in the GitHub actions we use [this shallow CI server image](https://github.com/nextcloud/docker-ci/pkgs/container/continuous-integration-shallow-server).
+Due to technical constraints the GitHub actions rely on a bare metal installation of Nextcloud on the runner in difference to local tests which can leverage much cleaner and reliable Docker containers.
 
-```sh
-#!/usr/bin/env zsh
+Currently, you still need to launch the server manually once before running any tests locally.
+This project comes with a convenience script to do so, see or run [Tests/Server.sh](Tests/Server.sh).
+**Please familiarize yourself with its content before running it** to avoid unintended removal of a container named like in the script. 
 
-# The fixed Docker container name for the Nextcloud server.
-CONTAINER_NAME="xcode-ui-test-server"
-
-# Check for environment variable of server version to use including a default value.
-SERVER_VERSION=${SERVER_VERSION:-"stable30"}
-
-# Launch Nextcloud Server in Docker Container.
-docker run \
-    --detach \
-    --name $CONTAINER_NAME \
-    --publish 8080:80 \
-    ghcr.io/nextcloud/continuous-integration-shallow-server:latest
-
-# Enable File Download Limit App.
-docker exec $CONTAINER_NAME su www-data -c "git clone --depth 1 -b $SERVER_VERSION https://github.com/nextcloud/files_downloadlimit.git /var/www/html/apps/files_downloadlimit/"
-docker exec $CONTAINER_NAME su www-data -c "php /var/www/html/occ app:enable files_downloadlimit"
-```
-
-As of now, you need to launch it manually once before running any tests.
-The instance is reused for all unit tests.
-Though, a long term goal is to automatically create and discard container _per test_ to ensure maximum test isolation and a clean test environment.
+The same instance is reused for all tests.
+A long term goal is to automatically create and discard container _per test_ to ensure maximum test isolation and a clean test environment.
 Note that plain HTTP is used in this case to circumvent certificate issues and simplify the sign-in flow in automation.
 
 ### UI Tests
 
 UI tests also require a Nextcloud server to test against just like integration tests.
-**Important**: Additionally, UI tests assume your test device (regardless whether physical or Simulator) is set to US English.
+**Important**: Additionally, UI tests assume your test device (regardless whether physical or Simulator) is set to **US English**.
 This is required because in some parts of the user interface the automation has no other choice than to rely on localized texts for control.
