--- Tests/NextcloudIntegrationTests/BaseIntegrationXCTestCase.swift
@@ -1,29 +1,0 @@
-//
-//  BaseIntegrationXCTestCase.swift
-//  
-// 
-//  Created by Milen Pivchev on 20.06.23.
-//  Copyright © 2023 Milen Pivchev. All rights reserved.
-//
-//  Author: Milen Pivchev <milen.pivchev@nextcloud.com>
-//  This program is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU General Public License as published by
-//  the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
-//
-//  This program is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU General Public License for more details.
-//
-//  You should have received a copy of the GNU General Public License
-//  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-//
-
-class BaseIntegrationXCTestCase: BaseXCTestCase {
-    var randomInt: Int {
-        get {
-            return Int.random(in: 1000...Int.max)
-        }
-    }
-}
