+++ Tests/NextcloudIntegrationTests/Placeholder.swift
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: Nextcloud GmbH
+// SPDX-FileCopyrightText: 2025 Iva Horn
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+// There are no unit tests yet.
+// This is an empty placeholder file.
+// The compiler requires at least an empty file for every target.
