--- Tests/NextcloudUITests/XCUIElement+awaitOrFail.swift
@@ -1,37 +1,0 @@
-// SPDX-FileCopyrightText: Nextcloud GmbH
-// SPDX-FileCopyrightText: 2025 Iva Horn
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-import XCTest
-
-extension XCUIElement {
-    ///
-    /// Convenience method to wait for the existence of a `XCUIElement` for the default time defined in the test constants.
-    ///
-    /// This will cause a test failure, if the user interface element did not appear or exist within given timeout.
-    /// Use ``await`` in case you want to check for the appearance of an `XCUIElement` without considering absence a failure.
-    ///
-    /// Example usage within a test method implementation:
-    ///
-    /// ```swift
-    /// element.awaitOrFail()
-    /// ```
-    ///
-    /// It is important to set `continueAfterFailure` to `false` for test implementations.
-    ///
-    /// > Important: This is a candidate for outsourcing into a dedicated library, if not NextcloudKit.
-    ///
-    /// - Parameters:
-    ///     - timeout: The number of seconds to wait.
-    ///     - file: File of the call point.
-    ///     - line: Source code line of the call point within `file`.
-    ///
-    func awaitOrFail(timeout: Double = TestConstants.controlExistenceTimeout, file: StaticString = #file, line: UInt = #line) {
-        guard waitForExistence(timeout: timeout) else {
-            XCTFail("Expected element did not exist after \(timeout) seconds.", file: file, line: line)
-            return
-        }
-
-        return
-    }
-}
