+++ iOSClient/Menu/NCShare+Menu.swift
@@ -60,10 +60,6 @@ extension NCShare {
                         advancePermission.downloadLimit = .limited(limit: downloadLimit.limit, count: downloadLimit.count)
                     }
 
-                    if let downloadLimit = try? self.database.getDownloadLimit(byAccount: self.metadata.account, shareToken: share.token) {
-                        advancePermission.downloadLimit = .limited(limit: downloadLimit.limit, count: downloadLimit.count)
-                    }
-
                     navigationController.pushViewController(advancePermission, animated: true)
                 }
             )
